pixly-backend/requirements.txt
appnope==0.1.2
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
boto3==1.18.28
botocore==1.21.28
cffi==1.14.6
click==8.0.1
decorator==5.0.9
dnspython==2.1.0
flake8==3.9.2
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==3.2
ipython==7.25.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.9.0
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.20
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
